fix: submit quiz options

+++ src/Modules/Tours/Explorer.Tours.API/Public/Tourist/IQuizSubmissionService.cs
@@ -9,6 +9,6 @@ namespace Explorer.Tours.API.Public.Tourist
 {
     public interface IQuizSubmissionService
     {
-        QuizResultDto SubmitAnswers(int quizId, QuizSubmissionDto submission);
+        QuizResultDto SubmitAnswers(long quizId, QuizSubmissionDto submission);
     }
 }
